refactor: move useGetUserQuery to data-provider and update imports
