+++ client/src/data-provider/Auth/index.ts
@@ -1,1 +1,2 @@
+export * from './queries';
 export * from './mutations';
