+++ client/src/hooks/AuthContext.tsx
@@ -10,9 +10,14 @@ import {
 import { useRecoilState } from 'recoil';
 import { useNavigate } from 'react-router-dom';
 import { setTokenHeader, SystemRoles } from 'librechat-data-provider';
-import { useGetUserQuery, useRefreshTokenMutation } from 'librechat-data-provider/react-query';
+import { useRefreshTokenMutation } from 'librechat-data-provider/react-query';
 import type { TLoginResponse, TLoginUser } from 'librechat-data-provider';
-import { useLoginUserMutation, useLogoutUserMutation, useGetRole } from '~/data-provider';
+import {
+  useLoginUserMutation,
+  useLogoutUserMutation,
+  useGetRole,
+  useGetUserQuery,
+} from '~/data-provider';
 import { TAuthConfig, TUserContext, TAuthContext, TResError } from '~/common';
 import useTimeout from './useTimeout';
 import store from '~/store';
